agregado search

+++ app/src/main/java/com/example/souvenirscadiz/data/model/Souvenir.kt
@@ -4,5 +4,5 @@ data class Souvenir(
     var nombre:String = "",
     var url:Int = 0,
     var tipo:Tipo = Tipo.PREDETERMINADO,
-    var precio:Int = 0
+    var precio:Double = 0.0
 )
